toolkit/content/tests/widgets/test_moz_checkbox.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>moz-checkbox tests</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script type="module" src="chrome://global/content/elements/moz-checkbox.mjs"></script>
</head>
<body>
<p id="display"></p>
<div id="content" style="max-width: fit-content;color: black;">
  <moz-checkbox id="moz-checkbox-1" label="test label 1" value="one"></moz-checkbox>
  <moz-checkbox id="moz-checkbox-2" label="test label 2" checked="" name="checkbox-2" value="two"></moz-checkbox>
  <moz-checkbox id="moz-checkbox-3" label="test label 3" disabled value="three"></moz-checkbox>
  <moz-checkbox id="moz-checkbox-4" label="test label 4" description="description"></moz-checkbox>
  <moz-checkbox id="moz-checkbox-5" label="test label 5">
    <span slot="description">slot description</span>
  </moz-checkbox>
  <moz-checkbox
  id="moz-checkbox-6"
  label="test label 6"
  description="description attr">
    <span slot="description">slot description</span>
  </moz-checkbox>
</div>
<pre id="test">
  <script class="testbody" type="application/javascript">
    const { TestUtils } = ChromeUtils.importESModule(
      "resource://testing-common/TestUtils.sys.mjs");
    const { BrowserTestUtils } = ChromeUtils.importESModule(
      "resource://testing-common/BrowserTestUtils.sys.mjs"
    );

    add_task(async function testMozCheckboxDisplay() {

      // Ensure moz-checkbox rendered correctly
      const mozCheckbox1 = document.getElementById("moz-checkbox-1");
      const mozCheckbox2 = document.getElementById("moz-checkbox-2");
      const mozCheckbox3 = document.getElementById("moz-checkbox-3");
      const mozCheckbox4 = document.getElementById("moz-checkbox-4");
      const mozCheckbox5 = document.getElementById("moz-checkbox-5");
      const mozCheckbox6 = document.getElementById("moz-checkbox-6");
      ok(mozCheckbox1, "moz-checkbox-1 is rendered");
      ok(mozCheckbox2, "moz-checkbox-2 is rendered");
      ok(mozCheckbox3, "moz-checkbox-3 is rendered");
      ok(mozCheckbox4, "moz-checkbox-4 is rendered");
      ok(mozCheckbox5, "moz-checkbox-5 is rendered");
      ok(mozCheckbox6, "moz-checkbox-6 is rendered");

      // Ensure label is set on moz-checkbox and inner label
      is(
        mozCheckbox1.labelEl.innerText,
        "test label 1",
        "inner label of moz-checkbox-1 is correct"
      );
      is(
        mozCheckbox2.labelEl.innerText,
        "test label 2",
        "inner label of moz-checkbox-2 is correct"
      );
      is(mozCheckbox3.labelEl.innerText, "test label 3", "inner label of moz-checkbox-3 is correct");
      is(mozCheckbox4.labelEl.innerText, "test label 4", "inner label of moz-checkbox-4 is correct");
      is(mozCheckbox5.labelEl.innerText, "test label 5", "inner label of moz-checkbox-5 is correct");
      is(mozCheckbox6.labelEl.innerText, "test label 6", "inner label of moz-checkbox-4 is correct");

      // Ensure checked attribute is set on moz-checkbox and inner checkbox
      ok(
        !mozCheckbox1.checked,
        "moz-checkbox-1 should not be checked on initial render"
      );
      ok(
        mozCheckbox2.checked,
        "moz-checkbox-2 should be checked on initial render"
      );
      ok(
        !mozCheckbox1.checkboxEl.checked,
        "inner checkbox of moz-checkbox-1 should not be checked on initial render");
      ok(
        mozCheckbox2.checkboxEl.checked,
        "inner checkbox of moz-checkbox-2 should be checked on initial render");
      ok(
        !mozCheckbox3.checked,
        "moz-checkbox-3 should not be checked on initial render"
      );
      ok(
        !mozCheckbox3.checkboxEl.checked,
        "inner checkbox of moz-checkbox-3 should not be checked on initial render"
      );
      ok(
        !mozCheckbox4.checked,
        "moz-checkbox-4 should not be checked on initial render"
      );
      ok(
        !mozCheckbox4.checkboxEl.checked,
        "inner checkbox of moz-checkbox-4 should not be checked on initial render"
      );
      ok(
        !mozCheckbox5.checked,
        "moz-checkbox-5 should not be checked on initial render"
      );
      ok(
        !mozCheckbox5.checkboxEl.checked,
        "inner checkbox of moz-checkbox-5 should not be checked on initial render"
      );
      ok(
        !mozCheckbox6.checked,
        "moz-checkbox-6 should not be checked on initial render"
      );
      ok(
        !mozCheckbox6.checkboxEl.checked,
        "inner checkbox of moz-checkbox-6 should not be checked on initial render"
      );

      // Ensure name propagates
      is(mozCheckbox2.checkboxEl.name, "checkbox-2", "name gets set on input");

      // Ensure value propagates
      is(mozCheckbox1.checkboxEl.value, "one", "value gets set on input 1");
      is(mozCheckbox2.checkboxEl.value, "two", "value gets set on input 2");
      is(mozCheckbox3.checkboxEl.value, "three", "value gets set on input 3");

      // Ensure disabled attribute is set on moz-checkbox and inner checkbox
      ok(
        !mozCheckbox1.disabled,
        "moz-checkbox-1 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox2.disabled,
        "moz-checkbox-2 should not be disabled on initial render"
      );
      ok(
        mozCheckbox3.disabled,
        "moz-checkbox-3 should be disabled on initial render"
      );
      ok(
        !mozCheckbox4.disabled,
        "moz-checkbox-4 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox5.disabled,
        "moz-checkbox-5 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox6.disabled,
        "moz-checkbox-6 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox1.checkboxEl.disabled,
        "inner checkbox of moz-checkbox-1 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox2.checkboxEl.disabled,
        "inner checkbox of moz-checkbox-2 should not be disabled on initial render"
      );
      ok(
        mozCheckbox3.checkboxEl.disabled,
        "inner checkbox of moz-checkbox-3 should be disabled on initial render"
      );
      ok(
        !mozCheckbox4.checkboxEl.disabled,
        "inner checkbox of moz-checkbox-4 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox5.checkboxEl.disabled,
        "inner checkbox of moz-checkbox-5 should not be disabled on initial render"
      );
      ok(
        !mozCheckbox6.checkboxEl.disabled,
        "inner checkbox of moz-checkbox-6 should not be disabled on initial render"
      );
    });

    add_task(async function testCheckboxEvents() {
      const mozCheckbox1 = document.getElementById("moz-checkbox-1");
      const mozCheckbox2 = document.getElementById("moz-checkbox-2");
      const mozCheckbox3 = document.getElementById("moz-checkbox-3");
      const mozCheckbox4 = document.getElementById("moz-checkbox-4");
      let checkboxes = [mozCheckbox1, mozCheckbox2, mozCheckbox3, mozCheckbox4];

      mozCheckbox1.checked = false;
      mozCheckbox2.checked = true;
      mozCheckbox3.disabled = true;
      mozCheckbox4.checked = false;
      await TestUtils.waitForTick();

      let seenEvents = [];
      function trackEvent(event) {
        seenEvents.push({
          type: event.type,
          checked: event.target.checked,
          localName: event.currentTarget.localName,
        });
      }

      function verifyEvents(expectedEvents) {
        is(
          seenEvents.length,
          expectedEvents.length,
          "moz-checkbox elements emit the expected number of events."
        );
        expectedEvents.forEach((eventInfo, index) => {
          let seenEventInfo = seenEvents[index];
          is(seenEventInfo.type, eventInfo.type, "Event type is correct.");
          is(
            seenEventInfo.localName,
            eventInfo.localName,
            "Event is emitted from the correct element"
          );
          is(
            seenEventInfo.checked,
            eventInfo.checked,
            "Event checked state is correct."
          );
        });
      }

      checkboxes.forEach(checkbox => {
        checkbox.addEventListener("click", trackEvent);
        checkbox.addEventListener("input", trackEvent);
        checkbox.addEventListener("change", trackEvent);
      })

      // Ensure that clicking the inner checkbox element emits the expected
      // events in the correct order.
      synthesizeMouseAtCenter(mozCheckbox1.checkboxEl, {});
      await TestUtils.waitForTick();

      verifyEvents([
        { type: "click", localName: "moz-checkbox", checked: true },
        { type: "input", localName: "moz-checkbox", checked: true },
        { type: "change", localName: "moz-checkbox", checked: true },
      ]);

      ok(
        mozCheckbox1.checked,
        "Clicking the inner checkbox should toggle the checked attribute"
      );

      // Reset seen events.
      seenEvents = [];

      // Ensure that clicking the inner label element emits the
      // expected events in the correct order.
      synthesizeMouseAtCenter(mozCheckbox1.labelEl, {});
      await TestUtils.waitForTick();

      // Since we click the label element, there is an additional
      // click event compared to the checkbox element, and this
      // first click doesn't update the checked value

      verifyEvents([
        { type: "click", localName: "moz-checkbox", checked: true },
        { type: "click", localName: "moz-checkbox", checked: false },
        { type: "input", localName: "moz-checkbox", checked: false },
        { type: "change", localName: "moz-checkbox", checked: false },
      ]);

      ok(
        !mozCheckbox1.checked,
        "Clicking the checkbox should toggle the checked attribute"
      );

      // Reset seen events.
      seenEvents = [];

      // Ensure that using the keyboard to activate the inner checkbox
      // emits the expected events in the correct order.

      mozCheckbox1.focus();
      synthesizeKey(" ", {});
      await TestUtils.waitForTick();

      verifyEvents([
        { type: "click", localName: "moz-checkbox", checked: true },
        { type: "input", localName: "moz-checkbox", checked: true },
        { type: "change", localName: "moz-checkbox", checked: true },
      ]);

      ok(
        mozCheckbox1.checked,
        "Activating the Space key on the inner checkbox should toggle the checked attribute"
      );

      // Ensure click() toggle the checkbox.
      mozCheckbox1.click();
      ok(!mozCheckbox1.checked, "click() toggled the checkbox");

      // Reset seen events.
      seenEvents = [];

      // Ensure clicking on a disabled moz-checkbox does not send
      // any events.

      synthesizeMouseAtCenter(mozCheckbox3.checkboxEl, {});
      await TestUtils.waitForTick();

      verifyEvents([]);

      // Ensure clicking on a description within moz-checkbox does not
      // change the value of the checkbox input.

      // Reset seen events.
      seenEvents = [];

      synthesizeMouseAtCenter(mozCheckbox4.descriptionEl, {});
      await TestUtils.waitForTick();

      verifyEvents([
        { type: "click", localName: "moz-checkbox", checked: false }
      ]);
      is(
        mozCheckbox4.checked,
        false,
        "Checkbox input should not change when clicking on description"
      );

    });

    // Verify setting iconSrc displays an icon in the checkbox.
    add_task(async function testCheckboxIcon() {
      const ICON_SRC = "chrome://global/skin/icons/edit-copy.svg";
      const mozCheckbox1 = document.getElementById("moz-checkbox-1");

      ok(!mozCheckbox1.icon, "No icon element present if iconSrc isn't set");
      mozCheckbox1.iconSrc = ICON_SRC;
      await mozCheckbox1.updateComplete;
      ok(mozCheckbox1.icon, "Checkbox displays an icon");
      is(
        mozCheckbox1.icon.getAttribute("src"),
        ICON_SRC,
        "Checkbox icon uses the expected source"
      );

      mozCheckbox1.iconSrc = null;
      await mozCheckbox1.updateComplete;
      ok(!mozCheckbox1.icon, "Checkbox icon can be removed");
    });

    // Ensure the description is generated correctly in the component.
    add_task(async function testCheckboxDescription() {
      const mozCheckbox4 = document.getElementById("moz-checkbox-4");
      let descriptionEl = mozCheckbox4.descriptionEl
      ok(descriptionEl, "Description element should exist");
      is(
        descriptionEl.innerText,
        "description",
        "The description element markup should have the correct text"
      );
      is(
        mozCheckbox4.checkboxEl.getAttribute("aria-describedby"),
        "description",
        "The input element should be described by the 'description' slot"
      );

      const mozCheckbox5 = document.getElementById("moz-checkbox-5");
      descriptionEl = mozCheckbox5.shadowRoot.querySelector("slot");

      ok(descriptionEl, "Description slot element should exist");
      is(descriptionEl.name, "description", "Description slow should have the correct name attribute");
      is(
        mozCheckbox5.querySelector("[slot='description']").innerText,
        "slot description",
        "The slotted element markup should have the correct text"
      );
      is(
        mozCheckbox5.checkboxEl.getAttribute("aria-describedby"),
        "description",
        "The input element should be described by the 'description' slot"
      );

      // Ensures that when given both a description via attribute and
      // slotted element, the description attribute takes precedence
      const mozCheckbox6 = document.getElementById("moz-checkbox-6");
      descriptionEl = mozCheckbox6.descriptionEl;
      ok(descriptionEl, "Description element should exist");
      is(
        descriptionEl.innerText,
        "description attr",
        "The description element markup should have the correct text"
      );
      is(
        mozCheckbox6.checkboxEl.getAttribute("aria-describedby"),
        "description",
        "The input element should be described by the 'description' slot"
      );

    });

  </script>
</pre>
</body>
</html>
